pyproject.toml
[tool.poetry]
name = "web13"
version = "0.1.0"
description = ""
authors = ["m1hanch <misha5264@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.24"
fastapi = "^0.108.0"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.37.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
jinja2 = "^3.1.2"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
uvicorn = "^0.25.0"
python-multipart = "^0.0.6"
bcrypt = "^4.1.2"
pytest = "^7.4.4"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"